refactor: change message and pending flush models to use i64 for IDs

+++ src/database/entities/messages.rs
@@ -6,10 +6,10 @@ use serenity::all::*;
 #[sea_orm(table_name = "messages")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub message_id: u64,
-    pub user_id: u64,
-    pub guild_id: u64,
-    pub channel_id: u64,
+    pub message_id: i64,
+    pub user_id: i64,
+    pub guild_id: i64,
+    pub channel_id: i64,
     pub timestamp: DateTime<Utc>,
 }
 
@@ -20,19 +20,19 @@ impl ActiveModelBehavior for ActiveModel {}
 
 impl Model {
     pub fn message_id(&self) -> MessageId {
-        MessageId::from(self.message_id)
+        MessageId::from(self.message_id as u64)
     }
 
     pub fn user_id(&self) -> UserId {
-        UserId::from(self.user_id)
+        UserId::from(self.user_id as u64)
     }
 
     pub fn guild_id(&self) -> GuildId {
-        GuildId::from(self.guild_id)
+        GuildId::from(self.guild_id as u64)
     }
 
     pub fn channel_id(&self) -> ChannelId {
-        ChannelId::from(self.channel_id)
+        ChannelId::from(self.channel_id as u64)
     }
 
     pub fn timestamp(&self) -> DateTime<Utc> {
